sim:epiphany: Improve performance and simplify handling of OOB events
Improve performance and simplify handling of out-of-band events.  The previous
commit that added initial OOB support introduced a performance hit of ~60%
(isns/s).  By only changing the PC in cgen if it was modified out-of-band, this
commit restores performance to ~85% compared to without OOB support.  Since we
only change the PC in cgen if it was modified OOB it also means we don't have
to increment it in 'epiphany_post_isn_callback'.