gdb: Fix unresolved symbols build failure
When building 'gdb' target, make sure user specified libraries in 'LDFLAGS' are
passed to linker after object files.

ESIM (part of epiphany sim target) depends on 'librt', which is not linked in
by default. The current workaround is to have the user call configure with:
'LDFLAGS=-lrt ./configure ...'
but for this to work on some platforms, e.g. Ubuntu 14.04, this patch is also
required.