sim:epiphany: Rework OOB handling (again)
The previous implementation had two limitiations that made it
unreliable.
- no causal ordering of register writes.
- in the case of two consecutive writes to same the register before the
  owning process sees the first write, important bits that would change
  the CPU state might be overwritten.

This implementation serializes all Special Core Register writes
on the owning process, which ensures both causality and that all writes
are seen by the owning process. Each CPU has a write slot protected by
a mutex. When a CPU wants to write to another CPUs SCR it must first
acquire the mutex and then wait for the write-slot to become available.