sim:epiphany:esim: Allow loads from remote GPRs when remote is inactive
Previously we just assumed the core was active and returned an error.
Instead, actually do check if the core is active and allow the load if it
isn't.

This is a little racy since epiphany_cpu_is_active() does not lock the
cpu state. But I think it should be fine since the manual says that
reads to a GPR of an active core is undefined.