cpu:epiphany: Change to unsigned type for registers
After changing to 64-bit addressing in the simulator a large number of GCC
tests started to fail. The cause was that for some instructions that
use register values for memory addressing, values with the most
significant bit set would be sign extended. E.g., a 0x80000000 register
value would translate to address 0xffffffff80000000 instead of
0x0000000080000000.

Change to unsigned data type for registers to stay clear of sign
extending negative values when upcasting to a larger word type.

cpu/
	* epiphany.cpu (h-all-registers) (h-registers) (h-core-registers)
	(h-coredma-registers) (h-coremem-registers)
	(h-coremesh-registers): Change to unsigned type.
	(h-fpregisters): Remove hardware.
	(h-registers-df): Likewise.
	(frd) (frn) (frm) (frd6) (frn6) (frm6) (rd6-df) (rn6-df)
	(rm6-df): Remove operands.
	(op-two_operands-float): Use double-integer operands for
	float64 instructions.
	(dnim fmax64) (dnim fmax64.l): Likewise.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>