bfd:epiphany: Fix software cache merge breakage
Fix merge from binutils-2.25 to binutils-2.27 issue that breaks
software caching.

In commit:
    "Always create dynamic sections for -E/--dynamic-list"
    commit bf89386a862ace008f0152bca8bddf996d3993c8
    Author: H.J. Lu <hjl.tools@gmail.com>
    Date:   Thu Feb 18 03:13:19 2016 -0800

Changes the behavior in
elf_link_add_object_symbols ()
to call _bfd_elf_link_create_dynamic_sections () if info->dynamic is
set.

We currently don't support that in the backend, and we don't need it
with the current software caching implementation.

So for now, set info->dynamic to 0 in epiphany_elf_check_relocs ().

Related commit (once we do implement it):

    "Set dynamic tag VMA and size from dynamic section when possible"
    commit 4ade44b727ee77adaa9c22719935d012e253a5e6
    Author: Alan Modra <amodra@gmail.com>
    Date:   Fri May 13 11:48:56 2016 +0930

bfd/
	* elf32-epiphany.c (epiphany_elf_create_plt_section): Don't set
	dynamic flag.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>