sim:epiphany: Fix issue with --e-core-mem 1M
In the previous commit:
esim: epiphany: Support passing command line options"
we use the cluster configuration to determine where to put the stack.

When a user provides a --e-core-mem value higher than the MMR base
(0xf0000), sim-if:epiphany_user_init could place the stack in the
registers area, or above the register area where access is not allowed.

To work around this, silently truncate core memory to be at most
ES_CORE_MMR_BASE.

sim/epiphany/
	* esim/esim.c (es_validate_cluster_cfg): Allow
	non power of two core_phys_mem.
	(es_init_impl): Truncate core_phys_mem to ES_CORE_MMR_BASE.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>