sim:epiphany: Don't adjust stack pointer if __stack is global address
In user environment mode we adjust the stack pointer so that we can use
all available core-local SRAM memory for passing the arguments to
"main". Only adjust the stack pointer if "__stack" points to a
core-local address. Why? Because if __stack is at a global address we
can assume that that there is more available memory there than there
would have been in local SRAM.

sim/epiphany/
	* sim-if.c (find_sym): New function.
	(epiphany_user_init): Do not move stack if __stack is a global
	address.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>