Header conflict resolved.
The header file demangle.h has changed in release 2.32 making it incompatible with the ones shipping with gcc and gdb.
For inclusion in the unified source this conflict causes build failure in binutils. Renaming demangle.h to demangle2.h allows for both revisions to coexist and building from unified source to succeed.